requirements.txt
librosa==0.9.2
matplotlib==3.6.2
numpy==1.23.0
pandas==1.5.2
Pillow==9.4.0
pydub==0.25.1
pytorch_lightning==1.8.4
scikit_learn==1.2.1
streamlit==1.16.0
torch==1.13.0
torchmetrics==0.11.0
torchvision==0.14.0
